chore: move locked wallet component
