+++ src/components/Layout/RequireUnlockedWallet.tsx
@@ -1,6 +1,6 @@
 import { FC } from "react";
 import { Navigate, Outlet, useParams } from "react-router-dom";
-import { useWalletStore } from "../store/wallet.store";
+import { useWalletStore } from "../../store/wallet.store";
 
 // Guard to prevent people from navigating to the wallet w/its id (or trying to)
 export const RequireUnlockedWallet: FC = () => {
